+++ braintree/exchange_rate_quote.py
@@ -1,0 +1,6 @@
+from braintree.attribute_getter import AttributeGetter
+from braintree.montary_amount import MontaryAmount
+
+class ExchangeRateQuote(AttributeGetter):
+    def __init__(self,attributes):
+        AttributeGetter.__init__(self,attributes)